Fixed the various bugs rdkit made with the conda environment. changed the dataset file to work with file object and not MolSupplier cause it runs faster, the reason is that smi file extention is like a text file and can just be read line by line.
Created a basic recurrent model that uses LSTM for SMILES generation, the model uses some arbitrary parans for now.
Added a basic trainning loop in the train.py file and a basic evaluate procedure which generates SMILES that are up to 36 characters, and anotehr method that calcualtes various stats from the metrics files in the utils module, right now it only computes the numebr of valid smiles in the train, and generated set and the QED kde plot of the train-set.

added various util methods to the utils file instaed of the stub it had previously.